#10 수정
수정할때도 이미지 삭제하는 로직 변경

+++ src/main/java/com/example/backend/board/service/BoardService.java
@@ -168,7 +168,9 @@ public class BoardService {
             board.addParticipatingCount();
         }
 
-        awsS3Service.deleteImage(board.getImageUrl().split(MsgEnum.IMAGE_DOMAIN.getMsg())[1]);
+        if(board.getImageUrl().split(MsgEnum.IMAGE_DOMAIN.getMsg()).length >= 2 ){
+            awsS3Service.deleteImage(board.getImageUrl().split(MsgEnum.IMAGE_DOMAIN.getMsg())[1]);
+        }
         board.update(requestDto.getBoard(), user);
 
     }
